+++ transports/quic/tests/stream_compliance.rs
@@ -1,7 +1,7 @@
 use futures::channel::oneshot;
 use futures::StreamExt;
-use libp2p_core::transport::ListenerId;
-use libp2p_core::Transport;
+use libp2p_core::transport::{DialOpts, ListenerId, PortUse};
+use libp2p_core::{Endpoint, Transport};
 use libp2p_quic as quic;
 use std::time::Duration;
 
@@ -47,7 +47,15 @@ async fn connected_peers() -> (quic::Connection, quic::Connection) {
             listener.next().await;
         }
     });
-    let dial_fut = dialer.dial(listen_address).unwrap();
+    let dial_fut = dialer
+        .dial(
+            listen_address,
+            DialOpts {
+                role: Endpoint::Dialer,
+                port_use: PortUse::Reuse,
+            },
+        )
+        .unwrap();
     async_std::task::spawn(async move {
         let connection = dial_fut.await.unwrap().1;