+++ protocols/perf/CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 0.4.0
+
+<!-- Update to libp2p-swarm v0.45.0 -->
 - Add ConnectionError to FromSwarm::ConnectionClosed.
   See [PR 5485](https://github.com/libp2p/rust-libp2p/pull/5485).