+++ transports/quic/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.11.1
+
+- Update `libp2p-tls` to version `0.5.0`, see [PR 5547]
+
+[PR 5547]: https://github.com/libp2p/rust-libp2p/pull/5547
+
+## 0.11.0
+
+- Implement refactored `Transport`.
+  See [PR 4568](https://github.com/libp2p/rust-libp2p/pull/4568)
+
 ## 0.10.3
 
 - Update `quinn` to 0.11 and `libp2p-tls` to 0.4.0.