+++ transports/webrtc-websys/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.4.0-alpha.2
+
+- Bump version of web-sys and update `__Nonexhaustive` to `__Invalid`.
+  See [PR 5569](https://github.com/libp2p/rust-libp2p/pull/5569) 
+
+## 0.4.0-alpha
+
+- Implement refactored `Transport`.
+  See [PR 4568](https://github.com/libp2p/rust-libp2p/pull/4568)
+
 ## 0.3.0-alpha
 
 - Bump version in order to publish a new version dependent on latest `libp2p-core`.